requirements.txt
stevedore
kombu
oslo.config
pbr
